Membuat CRUD di Produk

+++ admin/kategori-detail.php
@@ -1,6 +1,6 @@
 <?php
-    require "admin/session.php";
-    require "koneksi.php";
+    require "session.php";
+    require "../koneksi.php";
 
     $id = $_GET['p'];
 
@@ -13,7 +13,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Detail Kategori</title>
-    <link rel="stylesheet" href="bootstrap/css/bootstrap.min.css">
+    <link rel="stylesheet" href="../bootstrap/css/bootstrap.min.css">
 </head>
 <style>
     .form-control:focus {
@@ -78,6 +78,17 @@
                 }
 
                 if(isset($_POST['deleteBtn'])){
+                    $queryCheck = mysqli_query($con, "SELECT * FROM produk WHERE kategori_id='$id'");
+                    $dataCount = mysqli_num_rows($queryCheck);
+                    if($dataCount>0){
+                        ?>
+                            <div class="alert alert-primary mt-3" role="alert">
+                                Kategori tidak bisa dihapus karena sudah digunakan di produk
+                            </div>
+                        <?php 
+                        die();
+                    }
+
                     $queryDelete = mysqli_query($con, "DELETE FROM kategori WHERE id='$id'");
                     if($queryDelete){
                         ?>
@@ -93,6 +104,6 @@
             ?>
         </div>
     </div>
-    <script src="bootstrap/js/bootstrap.bundle.min.js"></script>
+    <script src="../bootstrap/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
